feat(table): deletion logic

+++ src/app/api/links/route.js
@@ -1,6 +1,7 @@
 import dbConnect from '../../../lib/dbConnect';
 import Link from '../../../models/link';
 import { verifyToken } from '../../../lib/auth';
+import Clicks from '@/models/click';
 
 export async function POST(request) {
     try {
@@ -53,24 +54,45 @@ export async function GET(request) {
 export async function DELETE(request) {
     try {
         await dbConnect();
-        const token = request.headers.get('Authorization')?.split(' ')[1];
+        const token = request.headers.get("Authorization")?.split(" ")[1];
         if (!token) {
-            return new Response(JSON.stringify({ error: 'Authentication required' }), { status: 401 });
+            return new Response(
+                JSON.stringify({ error: "Authentication required" }),
+                { status: 401 }
+            );
         }
         const decoded = verifyToken(token);
         const userId = decoded.userId;
 
-        // Get the shortHash from the query parameters
+        // Get the shortUrl and linkId from the query parameters
         const { searchParams } = new URL(request.url);
-        const shortUrl = searchParams.get('shortUrl');
-        if (!shortUrl) {
-            return new Response(JSON.stringify({ error: 'shortUrl is required' }), { status: 400 });
+        const shortUrl = searchParams.get("shortUrl");
+        const linkId = searchParams.get("linkId");
+
+        console.log('linkId', linkId)
+        console.log('shortUrl', shortUrl)
+
+        if (!shortUrl || !linkId) {
+            return new Response(
+                JSON.stringify({ error: "shortUrl and linkId are required" }),
+                { status: 400 }
+            );
         }
 
-        // Delete the link for the authenticated user and given shortHash
+        // Delete the link for the authenticated user and given shortUrl
         await Link.deleteOne({ userId, shortUrl });
-        return new Response(JSON.stringify({ message: 'Link deleted successfully' }), { status: 200 });
+
+        // Delete all clicks associated with the given linkId and userId
+        await Clicks.deleteMany({ linkId });
+
+        return new Response(
+            JSON.stringify({ message: "Link and its clicks deleted successfully" }),
+            { status: 200 }
+        );
     } catch (error) {
-        return new Response(JSON.stringify({ error: error.message }), { status: 500 });
+        return new Response(
+            JSON.stringify({ error: error.message }),
+            { status: 500 }
+        );
     }
 }
